ml-face-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1

pydantic==2.10.3
pydantic-settings==2.10.1

python-multipart
python-dotenv

numpy==1.26.4
pillow==11.0.0
opencv-python-headless

face-recognition==1.3.0
